backend/src/requirements.txt
faster-whisper==1.1.1
numpy==2.1.3
pydub==0.25.1
python-ffmpeg==2.0.12
torch==2.6.0+cu118
tqdm==4.67.1
typing_extensions==4.12.2
fastapi[standard]
typing
pathlib
pandas
bs4
datetime
os
uuid
requests
tempfile
io
openpyxl
ctranslate2>=3.22.0
werkzeug-3.1.3
sqlalchemy
database
psycopg2